Created the IDBCursorDirection enum for way the cursor traverses the set of data.
Added a default for the cursor to be in ascending order.
Updated the Query and QueryFromIndex to take the IDBCursorDirection as the last argument.
Added two new buttons to get the Employee Data in the descending order.
Created two new methods used by the new test buttons for retrieving in reverse order.